Add a wait online class to improve first-run
By default, the first run may fail to apply some resources as the
apiserver takes a few seconds before it opens its communication socket
when started.

This adds a simple exec that waits up to 30s for the apiserver to be
ready before any kubectl_apply is allowed to occur.